% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf_time_series.R
\name{coord_distance}
\alias{coord_distance}
\title{Return the distance (in km) between the latitude/longitude pairs
using the Haversine formula.}
\usage{
coord_distance(latA, lonA, latB, lonB)
}
\arguments{
\item{latA}{The latitude for the first coordinate pair, in
decimal-degrees Northing (or a vector of latitudes).}

\item{lonA}{The longitude for the first coordinate pair, in
decimal-degrees Easting (or a vector of longitudes).}

\item{latB}{The latitude for the second coordinate pair, in
decimal-degrees Northing (or a vector of latitudes).}

\item{lonB}{The longitude for the second coordinate pair, in
decimal-degrees Easting (or a vector of longitudes).}
}
\value{
The distance (in kilometres) between the coordinate pairs
(or a vector of respective distances).
}
\description{
Return the distance (in km) between the latitude/longitude pairs
using the Haversine formula.
}
\examples{
coord_distance(-34.9257, 138.5832, -34.9285, 138.6007)

}
