% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf_time_series.R
\name{extract_time_series}
\alias{extract_time_series}
\title{Return a data frame comprising the time series data for the
given NetCDF climate grid near the specified coordinates.}
\usage{
extract_time_series(nc_file, latitude, longitude, var = NULL)
}
\arguments{
\item{nc_file}{The file path for the NetCDF file.}

\item{latitude}{The latitude for the desired time series dataset
(in decimal-degrees Northing).}

\item{longitude}{The longitude for the desired time series dataset
(in decimal-degrees Easting).}

\item{var}{(Optional) The variable to extract the time series for
(default=NULL, in which case the first variable is extracted).}
}
\value{
A data frame containing the dates, nearest grid latitude and
longitude, and the variable values.
}
\description{
Return a data frame comprising the time series data for the
given NetCDF climate grid near the specified coordinates.
}
\examples{
\dontrun{
time_series <- extract_time_series("test.nc", -34.9257, 138.583)
}
}
