% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier.down}
\alias{outlier.down}
\title{Downweight outliers}
\usage{
outlier.down(data, model, cutoff = 3)
}
\arguments{
\item{cutoff}{}
}
\description{
Downweight outliers
}
\examples{
\dontrun{
JULES COMPLETE
}
}
