% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier.rem}
\alias{outlier.rem}
\title{Outlier removal function}
\usage{
outlier.rem(data, model, cutoff = 3)
}
\arguments{
\item{cutoff}{}
}
\description{
Outlier removal function
}
\examples{
\dontrun{
JULES COMPLETE
}
}
