% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.data.R
\name{pad.data}
\alias{pad.data}
\title{Title}
\usage{
pad.data(data, pattern = "Row:Column", split = "Block", keep = 4, fill = NULL)
}
\arguments{
\item{fill}{}
}
\description{
Title
}
\examples{
\dontrun{
JULES COMPLETE
}
}
