% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald.test}
\alias{wald.test}
\title{Wald tests for linear hypotheses}
\usage{
wald.test(object, ...)
}
\arguments{
\item{object}{An object. Methods are available for \code{asreml}.}

\item{...}{Passed to methods.}
}
\description{
Wald tests for linear hypotheses
}
