% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\title{Associate BLUEs with BLUPs from predict calls of the same model}
\usage{
associate(
  model,
  ran.term = "Treatment:Cultivar",
  fix.term = "Treatment:Type",
  ...
)
}
\arguments{
\item{model}{An \code{asreml} model.}

\item{ran.term}{Random terms from the \code{model}.}

\item{fix.term}{Fixed terms from the \code{model}.}

\item{...}{Other parameters to be passed to the predict function.}
}
\value{
JULES COMPLETE
}
\description{
Associate BLUEs with BLUPs from predict calls of the same model
}
\examples{
\dontrun{
JULES COMPLETE
}

}
