% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\title{Associate BLUEs with BLUPs form predict calls of the same model}
\usage{
associate(
  model,
  ran.term = "Treatment:Cultivar",
  fix.term = "Treatment:Type",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Associate BLUEs with BLUPs form predict calls of the same model
}
