% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiometryTools.R
\docType{package}
\name{BiometryTools}
\alias{BiometryTools}
\title{BiometryTools: A package containing some useful functions for linear mixed model analysis}
\description{
The BiometryTools package provides ...
}
\section{Foo functions}{

The foo functions ...
}

