% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{BLUEs LSD/p-value comparison function}
\usage{
compare(model, term = "Line", type = "PVAL", average.LSD = FALSE)
}
\arguments{
\item{average.LSD}{}
}
\value{

}
\description{
BLUEs LSD/p-value comparison function
}
