% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{BLUEs LSD/p-value comparison function}
\usage{
compare(model, term = "Line", type = "PVAL", average.LSD = FALSE)
}
\arguments{
\item{model}{An \code{asreml} model}

\item{term}{JULES COMPLETE}

\item{type}{The type of comparison. Can take values of \code{PVAL} or \code{LSD}.}

\item{average.LSD}{JULES COMPLETE}
}
\value{

}
\description{
BLUEs LSD/p-value comparison function
}
\examples{
\dontrun{
JULES COMPLETE
}
}
