% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv3.R
\name{conv}
\alias{conv}
\title{Conversion function for Efficiency and Responsiveness BLUPs in
Treatment x Site x Variety experiments}
\usage{
conv(model, Env = "TSite:Variety", levs = NULL, sep = "-", ...)
}
\arguments{
\item{model}{Final full Treatment x Site x Variety model.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Conversion function for Efficiency and Responsiveness BLUPs in
Treatment x Site x Variety experiments
}
