% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Title}
\usage{
extract(
  data,
  pattern = "Row:Column",
  match = "DH",
  split = "Block",
  pad = TRUE,
  keep = 4,
  fill = NULL
)
}
\arguments{
\item{fill}{}
}
\value{

}
\description{
Title
}
