% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{fast}
\alias{fast}
\title{FAST: overall performance and stability for interpreting Factor Analytic models}
\usage{
fast(model, dat = NULL, term = "fa(Site, 4):Genotype", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
FAST: overall performance and stability for interpreting Factor Analytic models
}
