% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineMap.R
\name{fineMap}
\alias{fineMap}
\title{Fine map a \code{wgaim} object}
\usage{
fineMap(
  model,
  intervalObj,
  mark = NULL,
  flanking = 50,
  exclusion.window = 10000,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fine map a \code{wgaim} object
}
\examples{
\dontrun{
JULES COMPLETE
}
}
