% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresion_functions.R
\name{fixedRegress}
\alias{fixedRegress}
\title{Fixed regression for doing stuff}
\usage{
fixedRegress(
  model,
  term = "Treatment:Genotype",
  levs = c("9 cm", "Control"),
  robust = TRUE
)
}
\arguments{
\item{robust}{}
}
\value{

}
\description{
Fixed regression for doing stuff
}
\examples{
\dontrun{
JULES COMPLETE
}
}
