% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_packages.R
\name{gh_packages}
\alias{gh_packages}
\title{Checks if a package was installed from GitHub}
\usage{
gh_packages(pkg)
}
\arguments{
\item{pkg}{A package to check}
}
\value{
A dataframe containing packages installed from remote sources, with the type of remote, account name and repo name.
}
\description{
Checks if a package was installed from GitHub
}
\keyword{internal}
