% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritability.R
\name{herit.asreml}
\alias{herit.asreml}
\title{Heritability for multi/single environment trials}
\usage{
herit.asreml(model, term = "SYear:Genotype", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Heritability for multi/single environment trials
}
