% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsd.R
\name{hsd}
\alias{hsd}
\title{Title}
\usage{
hsd(
  model,
  term = "Treatment:Genotype",
  by = "Treatment",
  omit.string = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{
JULES COMPLETE
}
}
