% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan plot using ggplot}
\usage{
manhattan(mlist, cross, chr.in = NULL, annotate = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Manhattan plot using ggplot
}
\examples{
\dontrun{
JULES COMPLETE
}
}
