% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clones.R
\name{phenClones}
\alias{phenClones}
\title{check whether genetic clones are in the experiment}
\usage{
phenClones(
  model,
  cross,
  matching = "Genotype",
  Envir = NULL,
  no.samp = 1000,
  sep = "_"
)
}
\arguments{
\item{sep}{}
}
\value{

}
\description{
check whether genetic clones are in the experiment
}
