% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clones.R
\name{phenClones}
\alias{phenClones}
\title{Check whether genetic clones are in the experiment}
\usage{
phenClones(
  model,
  cross,
  matching = "Genotype",
  Envir = NULL,
  no.samp = 1000,
  sep = "_"
)
}
\arguments{
\item{model}{An \code{asreml} model.}

\item{cross}{JULES COMPLETE}

\item{matching}{JULES COMPLETE}

\item{Envir}{JULES COMPLETE}

\item{no.samp}{JULES COMPLETE}

\item{sep}{JULES COMPLETE}
}
\value{
A data frame containing JULES COMPLETE
}
\description{
Check whether genetic clones are in the experiment
}
\examples{
\dontrun{
JULES COMPLETE
}
}
