% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clones.R
\name{phenfixClones}
\alias{phenfixClones}
\title{Fix clones if they are in the experiment}
\usage{
phenfixClones(data, cross, matching = "Genotype", sep = "_")
}
\arguments{
\item{data}{The data frame to fix}

\item{cross}{JULES COMPLETE}

\item{matching}{The column name to match on}

\item{sep}{The separator between JULES COMPLETE}
}
\value{
The JULES COMPLETE
}
\description{
Fix clones if they are in the experiment
}
\examples{
\dontrun{
JULES COMPLETE
}
}
