% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clones.R
\name{phenfixClones}
\alias{phenfixClones}
\title{Fix clones if they are in the experiment}
\usage{
phenfixClones(data, cross, matching = "Genotype", sep = "_")
}
\arguments{
\item{sep}{}
}
\value{

}
\description{
Fix clones if they are in the experiment
}
