% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresion_functions.R
\name{randomRegress}
\alias{randomRegress}
\title{Conversion function for Efficiency and Responsiveness BLUPs in Treatment x Site x Variety experiments}
\usage{
randomRegress(
  model,
  Env = "TSite:Variety",
  levs = NULL,
  sep = "-",
  pev = TRUE,
  ...
)
}
\arguments{
\item{model}{An \code{asreml} object. The final full Treatment x Site x Variety model}

\item{Env}{Treatment x Site x Variety term as a character.}

\item{levs}{Named treatment levels used in transformation. e.g. \code{c("Treat1", "Treat2")} would regress Treat2 on Treat1}

\item{sep}{separator used for Treat x Site names (if multi-x model), if not present assumes single section}

\item{...}{Other parameters passed to \code{\link[asreml:predict.asreml]{asreml::predict.asreml()}}.}
}
\description{
The function assumes you have a Treatment x Site factor that is a composite of treatments and sites. The function requires no specific ordering of the factor levels.
}
\examples{
\dontrun{
JULES COMPLETE
}
}
