% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_packages.R
\name{reinstall_packages}
\alias{reinstall_packages}
\title{Reinstall packages}
\usage{
reinstall_packages(location = .libPaths()[1], source = FALSE)
}
\arguments{
\item{location}{Location to check for installed packages and to reinstall to. Defaults to the first option in \code{.libPaths()}.}

\item{source}{Logical. Install packages from source that have later source versions than binaries? This should usually be FALSE.}
}
\value{

}
\description{
Reinstall packages
}
