% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise_sunset_times.R
\name{sunrise_time}
\alias{sunrise_time}
\title{Return the time of sunrise given the date and GPS coordinates.}
\usage{
sunrise_time(datetime, latitude, longitude)
}
\arguments{
\item{datetime}{A POSIXct object for the local date (and timezone).}

\item{latitude}{The decimal-degrees for the latitude (Northing).}

\item{longitude}{The decimal-degrees for the longitude (Easting).}
}
\value{
A POSIXct object for the local time of sunrise.
}
\description{
Return the time of sunrise given the date and GPS coordinates.
}
\examples{
library(lubridate)
date <- as_datetime("2020-01-01", tz = "Australia/Adelaide")
sunrise_time(date, -35.69167, 136.9650)
}
