% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise_sunset_times.R
\name{sunset_time}
\alias{sunset_time}
\title{Return the time of sunset given the date and GPS coordinates.}
\usage{
sunset_time(datetime, latitude, longitude)
}
\arguments{
\item{datetime}{A POSIXct object for the local date (and timezone).}

\item{latitude}{The decimal-degrees for the latitude (Northing).}

\item{longitude}{The decimal-degrees for the longitude (Easting).}
}
\value{
A POSIXct object for the local time of sunset.
}
\description{
Return the time of sunset given the date and GPS coordinates.
}
