% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\alias{theme_design}
\alias{theme_design_heat}
\alias{theme_scatter}
\alias{theme_barplot}
\title{Custom \code{ggplot2} themes}
\usage{
theme_design(base_size = 11, base_family = "")

theme_design_heat(base_size = 11, base_family = "")

theme_scatter(base_size = 11, base_family = "")

theme_barplot(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Plot font size, given in pts.}

\item{base_family}{Font family used for the plot.}
}
\value{







}
\description{
Themes that provide some custom styling for \code{ggplot2} plots.
}
