% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_packages.R
\name{transfer_packages}
\alias{transfer_packages}
\title{Easily reinstall all currently installed packages on another machine or version of R.}
\usage{
transfer_packages(
  library = .libPaths()[1],
  output = "online",
  expiry = "14d",
  filename = "transfer_packages",
  quiet = F
)
}
\arguments{
\item{library}{The location of the library on the current machine to copy.}

\item{output}{One of \code{online} (the default), \code{gist} or \code{local}. Saves a list of installed packages to the chosen location, and provides instructions on how to use this to (re)install the packages elsewhere.}

\item{expiry}{Expiry for online file store. Ignored for other save locations.}

\item{filename}{Filename for the local output file. Ignored if \code{output} is not set to \code{local}.}

\item{quiet}{Logical (default \code{FALSE}). Suppress output if \code{TRUE}.}
}
\value{

}
\description{
Easily reinstall all currently installed packages on another machine or version of R.
}
