% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald.test.asreml}
\alias{wald.test.asreml}
\title{Title}
\usage{
\method{wald.test}{asreml}(object, cc, keep.fac = TRUE)
}
\arguments{
\item{object}{An \code{asreml} object.}

\item{cc}{JULES COMPLETE}

\item{keep.fac}{Logical (default \code{TRUE}). Keep factors?}
}
\value{
A list containing the wald output.
}
\description{
Title
}
\examples{
\dontrun{
JULES COMPLETE
}
}
